Make nested helper class into a presenter
Previously we had a number of RuboCop issues due to the use of
instance variables inside a helper. It's surprising to find a
class inside a helper module. A presenter seems like the most
appropriate place to extract this class into.
